requirements.txt
# Streamlit app
streamlit

# AWS S3 support
boto3

# Document parsing
pdfplumber
python-docx
pymupdf  # for PDF annotations

# OCR.Space API
requests

# Embeddings and similarity search
sentence-transformers
faiss-cpu
numpy

# GPT integration
openai>=1.0.0

# Optional local dev utilities
python-dotenv  # optional: only used if st.secrets not available
